+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # CHANGELOG
 
-## 2.0.1
+## 2.2.0
+
+* fix: throw exception if there are two properties with the same name in the
+  same class hierarchy
+
+## 2.1.0
 
 * fix: make `FileRepository` implement `ResetInterface`
 * style: improve profiler UI
@@ -10,6 +15,8 @@
 * refactor: refactor directory hashing code to a separate class for reusability
 * fix: disable `PropertyLister` autoconfiguration
 * fix: compatibility with reconstitutor 2.1
+* fix(AtributesPropertyLister): throw exception if there are two properties with
+  the same name in the same class hierarchy
 
 ## 2.0.0